Moved test files to __test__ map. Improved tests
