+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "start": "node app.js",
     "dev": "nodemon app.js",
-    "test": "mocha ./src/**/*",
+    "test": "npm run service-tests & npm run integration-tests",
+    "service-tests": "jest ./src/__tests__/test/*",
+    "integration-tests": "mocha ./src/__tests__/integration/**/* --exit",
     "lint": "eslint .",
     "populate": "node ./src/db/populate.js",
     "live_populate": "node ./src/db/live_populate.js"
@@ -35,6 +37,7 @@
     "chai-http": "^4.3.0",
     "eslint": "^8.29.0",
     "eslint-config-prettier": "^8.5.0",
+    "jest": "^29.3.1",
     "mocha": "^10.2.0",
     "nodemon": "^2.0.20",
     "prettier": "^2.8.1"
