+++ src/__tests__/test/GraphService.spec.js
@@ -1,6 +1,5 @@
-const { SqliteDataContext } = require("../../../db/sqllite.client");
-const DashboardConfigService = require("../dashboardConfig.service");
-const GraphConfigService = require("../GraphConfig.service");
+const { SqliteDataContext } = require("../../db/sqllite.client");
+const GraphConfigService = require("../../services/dashboardconfig/GraphConfig.service");
 
 describe('Test graph retrieval', ()=>{
     var SqlDb;
